+++ src/com/company/EmployeeSolution.java
@@ -4689,7 +4689,7 @@ Output: [1,2,2,3,5,6]
         return cnt;
     }
 
-    //---------------------------------------------------------------------------------------------------
+    //28th Nov---------------------------------------------------------------------------------------------------
     public int pivotInteger(int n) {
         int total = Math.abs(n * (n + 1) / 2);
 
@@ -4718,50 +4718,6 @@ Output: [1,2,2,3,5,6]
 
         return tl.size() - ind;
 
-    }
-
-    class Solution {
-        public ListNode removeNodes(ListNode head) {
-
-            TreeMap<Integer, Integer> tm = new TreeMap<>();
-
-            List<Integer> nodes = new ArrayList<>();
-            ListNode temp = head;
-            while (temp != null) {
-                nodes.add(temp.val);
-                temp = temp.next;
-            }
-
-            Set<Integer> indexDeleted = new HashSet<>();
-
-            for (int i = nodes.size() - 1; i >= 0; i--) {
-                tm.put(nodes.get(i), tm.getOrDefault(nodes.get(i), 0) + 1);
-
-                if (tm.higherKey(nodes.get(i)) != null && tm.higherKey(nodes.get(i)) > nodes.get(i)) {
-                    indexDeleted.add(i);
-                }
-            }
-
-
-            ListNode ptr = new ListNode(-1);
-            ptr.next = head;
-
-            ListNode cur = head;
-            ListNode prev = ptr;
-
-            int curr = 0;
-
-            while (cur != null) {
-                if (indexDeleted.contains(curr++)) {
-                    prev.next = cur.next;
-                } else {
-                    prev = cur;
-                }
-                cur = cur.next;
-            }
-            return ptr.next;
-
-        }
     }
 
     public int numberOfCuts(int n) {
@@ -4805,7 +4761,6 @@ Output: [1,2,2,3,5,6]
 
         return ans;
     }
-
 
     public int bestClosingTime(String customers) {
