+++ src/com/company/LinkedListProblems.java
@@ -1,5 +1,7 @@
 package com.company;
 
+import java.util.*;
+
 public class LinkedListProblems {
     //Author: Anand
     /*
@@ -95,5 +97,47 @@ public class LinkedListProblems {
         ListNode nextNode = reverse(curr, k, len - k);
         head.next = nextNode;
         return prev;
+    }
+
+    public ListNode removeNodes(ListNode head) {
+
+        TreeMap<Integer, Integer> tm = new TreeMap<>();
+
+        List<Integer> nodes = new ArrayList<>();
+        ListNode temp = head;
+        while (temp != null) {
+            nodes.add(temp.val);
+            temp = temp.next;
+        }
+
+        Set<Integer> indexDeleted = new HashSet<>();
+
+        for (int i = nodes.size() - 1; i >= 0; i--) {
+            tm.put(nodes.get(i), tm.getOrDefault(nodes.get(i), 0) + 1);
+
+            if (tm.higherKey(nodes.get(i)) != null && tm.higherKey(nodes.get(i)) > nodes.get(i)) {
+                indexDeleted.add(i);
+            }
+        }
+
+
+        ListNode ptr = new ListNode(-1);
+        ptr.next = head;
+
+        ListNode cur = head;
+        ListNode prev = ptr;
+
+        int curr = 0;
+
+        while (cur != null) {
+            if (indexDeleted.contains(curr++)) {
+                prev.next = cur.next;
+            } else {
+                prev = cur;
+            }
+            cur = cur.next;
+        }
+        return ptr.next;
+
     }
 }
